Restructured AI folder
